Add real total, complete and percentage on getCustomerStatusCountGroupByTasks response

+++ src/main/java/com/bank_dki/be_dms/dto/CustomerStatusCountDto.java
@@ -10,8 +10,8 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CustomerStatusCountDto {
-    private int total;
-    private int completed;
+    private Long total;
+    private Long completed;
     private List<Categories> categories;
 
     @Data
